backend/requirements.txt
# Encryption
cryptography==42.0.8

# FastAPI & server
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.6

# Auth
PyJWT==2.8.0
bcrypt==4.1.3

# Rate limiting
slowapi==0.1.9

# HTTP
httpx==0.27.0

# Bot / ML dependencies (shared with bot/)
MetaTrader5==5.0.4424
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
yfinance==0.2.41

# Other
python-dotenv==1.0.1
